Implement tqdm; Hotfix download from deleted chats; prettify output

+++ main.py
@@ -1,5 +1,6 @@
 import os
 import shutil
+from tqdm import tqdm
 from telethon import TelegramClient, events, sync
 from telethon.tl.types import InputMessagesFilterDocument
 
@@ -19,32 +20,41 @@ for i, dialog in enumerate(dialogs):
 
 while True:
     try:
-        selected_dialog = int(input("Please enter dialog number: "))
+        selected_dialog = int(input("\nPlease enter dialog number: "))
         if not (1 <= selected_dialog <= len(dialogs)):
             raise ValueError
         break
     except ValueError:
         print("Entered value is incorrect. Please try again...")
 
+print()
+
 dialog = dialogs[selected_dialog - 1]
-
-directory_name = dialog.name if dialog.name else dialog.entity.id
-
-if not os.path.exists(directory_name):
-    os.makedirs(directory_name)
-else:
-    shutil.rmtree(directory_name)
-    os.makedirs(directory_name)
-
-print(f"\n{'👾' * 40}")
-at_least_one_media_file_present = False
 entity = dialog.entity
-for message in client.iter_messages(entity, filter=InputMessagesFilterDocument):
-    at_least_one_media_file_present = True
-    print(f"\n⬇️  Downloading {message.file.name[:65]}")
-    message.download_media(directory_name)
 
-if at_least_one_media_file_present:
-    print(f"\n✅ Successfully saved all media to `{directory_name}` folder.")
+messages = [
+    message
+    for message in client.iter_messages(entity, filter=InputMessagesFilterDocument)
+]
+
+if messages:
+    saved_filenames = list()
+    directory_name = dialog.name if dialog.name else str(dialog.entity.id)
+
+    if not os.path.exists(directory_name):
+        os.makedirs(directory_name)
+    else:
+        shutil.rmtree(directory_name)
+        os.makedirs(directory_name)
+
+    for message in tqdm(messages):
+        at_least_one_media_file_present = True
+        saved_filenames.append(message.file.name)
+        message.download_media(directory_name)
+
+    print(f"\n✅ Successfully following files to `{directory_name}` folder:")
+
+    for filename in saved_filenames:
+        print(filename)
 else:
     print(f"\n❌ Could not find any media files.")
